backend/requirements.txt
numpy==1.24.0
tensorflow==2.10.0
transformers==4.30.2
flask==2.1.3
werkzeug==2.1.2
flask-cors==3.0.10
tweepy==4.14.0
pandas==1.5.3
python-dotenv==0.19.0
